# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: CityLink
  version: 0.0.1
servers:
  - url: http://localhost:4000/api
    description: Dev server
paths:
...
  /user/info:
    get:
      summary: Получение информации о пользователе
      tags: 
        - User
      operationId: getUserInfo
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow    
    
      responses:
        '200':
          description: Успешное получение данных
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    description: id пользователя
                    type: string
                    example: 00002d49-dbc6-41cd-a8be-6e4dfb891482
                  firstName:
                    description: Имя пользователя
                    type: string
                    example: Иван
                  lastName:
                    description: Фамилия пользователя
                    type: string
                    example: Иванов
                  middleName:
                    description: Отчество пользователя
                    type: string
                    example: Иванович
                  email:
                    description: email пользователя
                    type: string
                    example: example@gmail.com                     
                  gender:
                    description: Пол пользователя
                    type: boolean
                    example: true  
                  documentNumber:
                    description: Паспортные данные пользователя
                    type: string
                    example: 1111222222
                  birthDate:
                    description: Дата рождения пользователя
                    type: string
                    example: 2024-01-02T14:35:43.000Z
                  favouriteCity:
                    description: Самый посещаемый город пользователем
                    type: string
                    example: Пятигорск
                  favouriteCityCount:
                    description: Количество посещений любимого города
                    type: number
                    example: 5
                  ticketsAmount:
                    description: Количество купленных билетов
                    type: number
                    example: 10
                  phoneNumber:
                    description: Номер телефона пользователя
                    type: string
                    example: 79881233223
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
          
        '401':
          description:  Ошибка авторизации
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/401"
  /user/validateDate:
    post:
      summary: Проверка валидности данных
      tags: 
        - User
      operationId: dataValidation
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                phoneNumber:
                  description: Номер телефона пользователя
                  type: string
                  example: 79881233223
                email:
                  description: email пользователя
                  type: string
                  example: example@gmail.com
              required:
                - phoneNumber
                - email
              
      responses:
        '200':
          description: Успешная проверка данных
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    description: id пользователя
                    type: string
                    example: Request completed
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
  /user/update:
    put:
      summary: Обновление любых данных пользователя
      description: gender boolean, favouriteCityCount и ticketsAmount number, birthDate в формате timestamp с типом данных string, остальное string 
      tags: 
        - User
      operationId: updateProfile
      
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow 
      
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                phoneNumber:
                  description: Номер телефона пользователя
                  type: string
                  example: 79881233223
                email:
                  description: email пользователя
                  type: string
                  example: example@gmail.com
              required:
                - phoneNumber
                - email
              
      responses:
        '200':
          description: Успешное обновление данных
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Request completed
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"        
  /user/getUsers:
    get:
      summary: Получение списка пользователей
      description: Будут приходить все данные, включая createdAt, кроме password. В responses лишь пример, чтобы избежать большого кода.
      tags: 
        - User
      operationId: getUsers
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow    
    
      responses:
        '200':
          description: Успешное получение списка
          content:
            'application/json':
              schema:
                type: object
                properties:
                  users:
                    description: Массив объектов
                    type: array
                    items:
                      type: object
                      properties:
                        firstName:
                          type: string
                        lastName:
                          type: string
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"
  /user/deleteUser:
    delete:
      summary: Удаление пользователя
      tags: 
        - User
      operationId: deleteUser
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
            description: Токен авторизации
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImViMzhmYzFkLTI4YWQtNDY1Mi05OWU5LWJjNjA5YTBkZjllYSIsImlhdCI6MTcwNTg0MzYzOSwiZXhwIjoxNzA2Mjc1NjM5fQ.Xm-cygFpoVrHMaNoiXamGnDb7_LDbGPubNnDWK6BWow
        - name: id
          in: query
          schema:
            type: string
            description: id пользователя
            example: 00002d49-dbc6-41cd-a8be-6e4dfb891482
    
      responses:
        '200':
          description: Успешное удаление пользователя
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Request completed
                        
        '400':
          description:  Ошибка в отправляемых данных
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/400"
                
        '500':
          description:  Ошибка на стороне сервера
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/500"      
...
components:
  schemas:
    400: 
      description:  Ошибка в отправляемых данных
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
    500:
      description:  Ошибка на стороне сервера
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения
    401: 
      description:  Ошибка авторизации
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Пример сообщения